Updated function signature to use long long int

+++ org.lflang/src/lib/core/platform/lf_C11_threads_support.c
@@ -129,7 +129,7 @@ int lf_cond_wait(_lf_cond_t* cond, _lf_mutex_t* mutex) {
  * @return 0 on success, LF_TIMEOUT on timeout, and platform-specific error
  *  number otherwise (see pthread_cond_timedwait).
  */
-int lf_cond_timedwait(_lf_cond_t* cond, _lf_mutex_t* mutex, int64_t absolute_time_ns) {
+int lf_cond_timedwait(_lf_cond_t* cond, _lf_mutex_t* mutex, long long int absolute_time_ns) {
     // Convert the absolute time to a timespec.
     // timespec is seconds and nanoseconds.
     struct timespec timespec_absolute_time
